feat(alert) Update ui for watchlist step (use same component as in the trends)
