+++ src/lib/ui/app/Alerts/categories/watchlist/watchlist-form-step/ui/index.svelte
@@ -2,16 +2,16 @@
   import type { TBaseSchema } from '../schema.js'
   import type { TAlertStep } from '$ui/app/Alerts/form-steps/index.svelte.js'
 
-  import Checkbox from '$ui/core/Checkbox/index.js'
+  import ListOfWatchlists from './ListOfWatchlists.svelte'
 
   type TProps = { step: TAlertStep<TBaseSchema> }
 
   let { step }: TProps = $props()
+
+  const selectedId = $derived(step.state.$$.target.watchlist_id)
 </script>
 
-<Checkbox
-  isActive={step.isValid.$}
-  onCheckedChange={(value) => {
-    step.state.$$.target.watchlist_id = value ? 669 : null
-  }}
-></Checkbox>
+<ListOfWatchlists
+  {selectedId}
+  onSelect={(watchlist) => (step.state.$$.target.watchlist_id = watchlist?.id ?? null)}
+/>
